+++ client/client.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	pb "bms/bmsproto"
-	model "bms/model"
+	"bms/utils"
 	"context"
 	"fmt"
 	"log"
@@ -19,7 +19,7 @@ const (
 func main() {
 
 	conn, err := grpc.Dial(address, grpc.WithTransportCredentials(insecure.NewCredentials()))
-	model.CheckError(err)
+	utils.CheckError(err)
 	defer conn.Close()
 
 	client := pb.NewBmsDatabaseCrudClient(conn)
@@ -29,7 +29,7 @@ func main() {
 
 	//creating new user
 	new_user, err := client.CreateUser(ctx, &pb.NewUser{UserName: "PST", Password: "123", Email: "pst@bc.in", PhoneNumber: "9123"})
-	model.CheckError(err)
+	utils.CheckError(err)
 
 	log.Printf("User Name: %v, Password: %v, Email: %v, PhoneNumber : %v", new_user.GetUserName(), new_user.GetPassword(), new_user.GetEmail(), new_user.GetPhoneNumber())
 
@@ -44,7 +44,7 @@ func main() {
 		ReleaseDate: "1-3-2022",
 		Status:      "Active",
 	})
-	model.CheckError(err)
+	utils.CheckError(err)
 
 	log.Printf("Movie Name: %v, Director: %v, Description: %v, Rating : %v, Language : %v, Genre : %v, Release date : %v, Status : %v", new_movie.GetMovieName(), new_movie.GetDirector(), new_movie.GetDescription(), new_movie.GetRating(), new_movie.GetLanguage(), new_movie.GetGenre(), new_movie.GetReleaseDate(), new_movie.GetStatus())
 
@@ -54,7 +54,7 @@ func main() {
 		ShowId: 1,
 		Amount: 150,
 	})
-	model.CheckError(err)
+	utils.CheckError(err)
 
 	log.Printf("User Id: %v, Show Id: %v, Amount: %v", new_booking.GetUserId(), new_booking.GetShowId(), new_booking.GetAmount())
 
@@ -63,7 +63,7 @@ func main() {
 		BookingId:  1,
 		SeatNumber: 2,
 	})
-	model.CheckError(err)
+	utils.CheckError(err)
 
 	log.Printf("Booking Id: %v, Seat Number %v", new_seat.GetBookingId(), new_seat.GetSeatNumber())
 
@@ -75,13 +75,13 @@ func main() {
 		BookingId: 2,
 		UserId:    1,
 	})
-	model.CheckError(err)
+	utils.CheckError(err)
 
 	log.Printf("Amount: %v, Mode: %v, Status: %v", new_payment.GetAmount(), new_payment.GetMode(), new_payment.GetStatus())
 
 	// Get all movies
 	AllMovies, err := client.GetMovies(ctx, &pb.EmptyMovie{})
-	model.CheckError(err)
+	utils.CheckError(err)
 
 	for _, movie := range AllMovies.Movies {
 		fmt.Println(movie.GetMovieName(), movie.GetDirector(), movie.GetDescription(), movie.GetRating(), movie.GetLanguage(), movie.GetGenre(), movie.GetReleaseDate(), movie.GetStatus())
@@ -92,7 +92,7 @@ func main() {
 		Genre:  "Action",
 		Rating: 8,
 	})
-	model.CheckError(err)
+	utils.CheckError(err)
 
 	for _, movie := range MovieByPreference.Movies {
 		fmt.Println(movie.GetMovieName(), movie.GetDirector(), movie.GetDescription(), movie.GetRating(), movie.GetLanguage(), movie.GetGenre(), movie.GetReleaseDate(), movie.GetStatus())
@@ -102,7 +102,7 @@ func main() {
 	ShowsList, err := client.GetListOfShowsByTheatre(ctx, &pb.Theatre{
 		TheatreId: 1,
 	})
-	model.CheckError(err)
+	utils.CheckError(err)
 
 	for _, show := range ShowsList.Shows {
 		fmt.Println(show.GetDate(), show.GetStartTime(), show.GetEndTime(), show.GetMovieId(), show.GetTheatreId())
@@ -112,7 +112,7 @@ func main() {
 	BookingsList, err := client.GetListOfBookingsByUser(ctx, &pb.User{
 		Id: 1,
 	})
-	model.CheckError(err)
+	utils.CheckError(err)
 
 	for _, booking := range BookingsList.Bookings {
 		fmt.Println(booking.GetUserId(), booking.GetShowId(), booking.GetAmount())
@@ -120,21 +120,21 @@ func main() {
 
 	// updating user info
 	updated_user, err := client.UpdateUser(ctx, &pb.User{Password: "newPwd", PhoneNumber: "456", UserName: "GST", Id: 3})
-	model.CheckError(err)
+	utils.CheckError(err)
 	fmt.Println(updated_user)
 
 	// updating movie status
 	updatedMovieStatus, err := client.UpdateMovieStatus(ctx, &pb.Movie{Id: 3, Status: "Active"})
-	model.CheckError(err)
+	utils.CheckError(err)
 	fmt.Println(updatedMovieStatus)
 
 	// updating show details
 	updatedShowDetails, err := client.UpdateShowDetails(ctx, &pb.Show{Id: 1, MovieId: 2})
-	model.CheckError(err)
+	utils.CheckError(err)
 	fmt.Println(updatedShowDetails)
 
 	//  deleting booking
 	cancelledBooking, err := client.CancelBooking(ctx, &pb.Booking{Id: 3})
-	model.CheckError(err)
+	utils.CheckError(err)
 	fmt.Println(cancelledBooking)
 }
