+++ model/startDb.go
@@ -1,29 +1,24 @@
 package model
 
 import (
+	"bms/utils"
 	"flag"
-	"fmt"
 	"os"
 
 	"github.com/jinzhu/gorm"
-	"github.com/joho/godotenv"
 	_ "github.com/lib/pq"
 )
 
 func StartDB() (*gorm.DB, error) {
 
-	envErr := godotenv.Load(".env")
-	if envErr != nil {
-		fmt.Printf("Could not load .env file")
-		os.Exit(1)
-	}
+	utils.LoadEnv() // To load .env file
 
 	db_user := flag.String("user", "postgres", "database user")
 	db_password := os.Getenv("db_password")
 	db_name := os.Getenv("db_name")
 	conn := "user=" + *db_user + " password=" + db_password + " dbname=" + db_name + " sslmode=disable"
 	db, err := gorm.Open("postgres", conn)
-	CheckError(err)
+	utils.PanicError(err)
 
 	db.AutoMigrate(&User{})
 	db.AutoMigrate(&Booking{})
@@ -41,10 +36,4 @@ func StartDB() (*gorm.DB, error) {
 	db.Model(&Show{}).AddForeignKey("theatre_id", "theatres(id)", "CASCADE", "RESTRICT")
 
 	return db, nil
-}
-
-func CheckError(err error) {
-	if err != nil {
-		panic(err)
-	}
 }
